requirements.txt
pandas
requests
matplotlib
seaborn
pyyaml
